src/requirements.txt
argcomplete==1.9.3
asn1crypto==0.24.0
awscli==1.16.168
boto3==1.9.157
botocore==1.12.158
certifi==2019.3.9
cffi==1.12.3
cfn-flip==1.2.0
chardet==3.0.4
Click==7.0
colorama==0.3.9
cryptography==2.6.1
defusedxml==0.6.0
Django==2.1.8
django-allauth==0.39.1
django-cors-headers==2.5.2
django-rest-knox==4.0.1
djangorestframework==3.9.2
docutils==0.14
durationpy==0.5
future==0.16.0
hjson==3.0.1
idna==2.8
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
Markdown==3.1
oauthlib==3.0.1
placebo==0.9.0
pyasn1==0.4.5
pycparser==2.19
python-dateutil==2.6.1
python-slugify==1.2.4
python3-openid==3.1.0
pytz==2019.1
PyYAML==3.13
requests==2.21.0
requests-oauthlib==1.2.0
rsa==3.4.2
s3transfer==0.2.0
six==1.12.0
sqlparse==0.3.0
stripe==2.28.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.4.7
Unidecode==1.0.23
urllib3==1.24.3
Werkzeug==0.15.4
whitenoise==4.1.2
wsgi-request-logger==0.4.6
zappa==0.48.2
